# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SN Auto Parts API
  description: |
    Feature-config driven e-commerce API for auto parts.
    
    ## Authentication
    All authenticated endpoints require a session token passed via:
    - Cookie: `auth-token=<token>`
    - Header: `Authorization: Bearer <token>`
    
    ## Authorization
    APIs are generic - authorization is determined by the authenticated user's role 
    and feature configuration. The same endpoint may return different data or 
    permissions based on the user's role.
    
    ## Feature-Based Access
    Each endpoint requires specific features to be enabled in the user's role configuration.
    If a user lacks the required feature, they receive a 403 Forbidden response.
  version: 1.0.0
  contact:
    name: SN Auto Parts
    email: support@snautoparts.com
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.snautoparts.com/api/v1
    description: Production server
...
paths:
...
  /settings:
    get:
      tags: [Settings]
      summary: Get all settings
      operationId: getSettings
      security:
        - sessionAuth: []
      x-feature-required: settings.view
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Settings
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/DataWrapper'
                  - properties:
                      data:
                        type: object
                        additionalProperties: true
  /settings/{key}:
    get:
      tags: [Settings]
      summary: Get setting by key
      operationId: getSetting
      security:
        - sessionAuth: []
      x-feature-required: settings.view
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Setting value
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/DataWrapper'
                  - properties:
                      data:
                        $ref: '#/components/schemas/Setting'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Settings]
      summary: Update setting
      operationId: updateSetting
      security:
        - sessionAuth: []
      x-feature-required: settings.update
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties:
                value: {}
                category:
                  type: string
      responses:
        '200':
          description: Setting updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/DataWrapper'
                  - properties:
                      data:
                        $ref: '#/components/schemas/Setting'
  /settings/bulk:
    post:
      tags: [Settings]
      summary: Bulk update settings
      operationId: bulkUpdateSettings
      security:
        - sessionAuth: []
      x-feature-required: settings.update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Settings updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/DataWrapper'
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Setting'

  # ============================================
  # AUDIT ENDPOINTS
  # ============================================
...
  /reports/sales:
    get:
      tags: [Reports]
      summary: Get sales report
      operationId: getSalesReport
      security:
        - sessionAuth: []
      x-feature-required: reports.salesByDay
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [day, week, month]
            default: day
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Sales data
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/DataWrapper'
                  - properties:
                      data:
                        $ref: '#/components/schemas/SalesReport'
  /reports/categories:
    get:
      tags: [Reports]
      summary: Get sales by category
      operationId: getCategoryReport
      security:
        - sessionAuth: []
      x-feature-required: reports.salesByCategory
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Category sales data
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/DataWrapper'
                  - properties:
                      data:
                        $ref: '#/components/schemas/CategoryReport'
  /reports/inventory-value:
    get:
      tags: [Reports]
      summary: Get inventory value report
      operationId: getInventoryValueReport
      security:
        - sessionAuth: []
      x-feature-required: reports.inventoryValue
      responses:
        '200':
          description: Inventory value data
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/DataWrapper'
                  - properties:
                      data:
                        $ref: '#/components/schemas/InventoryValueReport'
  /reports/gmv:
    get:
      tags: [Reports]
      summary: Get GMV report
      operationId: getGmvReport
      security:
        - sessionAuth: []
      x-feature-required: reports.gmv
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [day, week, month, year]
            default: month
      responses:
        '200':
          description: GMV data
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/DataWrapper'
                  - properties:
                      data:
                        $ref: '#/components/schemas/GmvReport'

# ============================================
# COMPONENTS
# ============================================
components:
  securitySchemes:
    sessionAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Session token (also accepted as auth-token cookie)
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    DataWrapper:
      type: object
      properties:
        data: {}
...
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
        error:
          type: string
        message:
          type: string
        details:
          type: object

    # Auth Schemas
...
    Setting:
      type: object
      properties:
        id:
          type: string
        key:
          type: string
        value: {}
        category:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # Audit Schemas
...
    SalesReport:
      type: object
      properties:
        period:
          type: string
        totalSales:
          type: number
        orderCount:
          type: integer
        averageOrderValue:
          type: number
        data:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              sales:
                type: number
              orders:
                type: integer
    CategoryReport:
      type: object
      properties:
        totalSales:
          type: number
        data:
          type: array
          items:
            type: object
            properties:
              categoryId:
                type: string
              categoryName:
                type: string
              sales:
                type: number
              percentage:
                type: number
    InventoryValueReport:
      type: object
      properties:
        totalValue:
          type: number
        totalCost:
          type: number
        potentialProfit:
          type: number
        itemCount:
          type: integer
        lowStockCount:
          type: integer
        outOfStockCount:
          type: integer
        byCategory:
          type: array
          items:
            type: object
            properties:
              categoryName:
                type: string
              value:
                type: number
              itemCount:
                type: integer
    GmvReport:
      type: object
      properties:
        totalGmv:
          type: number
        inventoryGmv:
          type: number
        dropshipGmv:
          type: number
        orderCount:
          type: integer
        period:
          type: string
        data:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              gmv:
                type: number
              inventoryGmv:
                type: number
              dropshipGmv:
                type: number